image-merger/requirements.txt
Pillow==9.0.0
reportlab==3.6.1
svglib==1.1.0
PyMuPDF==1.19.6